выносит стор моковый в файл, рефакторит App снепшот тест
