+++ src/test-data/test-data.js
@@ -1,8 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-import {NameSpace} from "../store/reducers/root-reducer";
-import {AuthorizationStatus} from "../const";
-import configureStore from "redux-mock-store";
 
 export const noop = () => {};
 
@@ -30,21 +27,6 @@ MockComponent.propTypes = {
 };
 
 export {MockComponent};
-
-const configuredStore = configureStore([]);
-const mockedStore = configuredStore({
-  [NameSpace.DATA]: {
-    activeGenre: `All genres`,
-    films: filmListMock,
-    shownFilmsCount: 8,
-    isLoading: false,
-  },
-  [NameSpace.USER]: {
-    authorizationStatus: AuthorizationStatus.AUTH,
-  }
-});
-
-export {mockedStore};
 
 export const filmListMock = [
   {
